Move anago subcommands to own directory
We should stick to a naming convention which makes it easy to
distinguish "throw away" anago subcommands vs first-class krel
ones.
