+++ cmd/krel/cmd/root.go
@@ -23,6 +23,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	"k8s.io/release/cmd/krel/cmd/anago"
 	"k8s.io/release/pkg/log"
 )
 
@@ -68,6 +69,8 @@ func init() {
 	rootCmd.PersistentFlags().BoolVar(&rootOpts.cleanup, "cleanup", false, "cleanup flag")
 	rootCmd.PersistentFlags().StringVar(&rootOpts.repoPath, "repo", filepath.Join(os.TempDir(), "k8s"), "the local path to the repository to be used")
 	rootCmd.PersistentFlags().StringVar(&rootOpts.logLevel, "log-level", "info", "the logging verbosity, either 'panic', 'fatal', 'error', 'warn', 'warning', 'info', 'debug' or 'trace'")
+
+	rootCmd.AddCommand(anago.AnagoCmd)
 }
 
 func initLogging(*cobra.Command, []string) error {
